requirements.txt
onnxruntime
fastapi
pydantic
numpy
opencv-python-headless==4.2.0.32
pillow